+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## ChangeLog
 
 * Fixed entity fx force death doesn't work
+* Fixed shallow copy doesn't copy transform